make cucumber work - albeit jumping through hoops

+++ features/support/env.rb
@@ -7,6 +7,10 @@ require 'steam'
 
 require File.expand_path(File.dirname(__FILE__) + '/../../config/environment')
 
+Dir[Rails.root + 'app/models/**/*.rb'].each { |f| require f }
+
+require 'rspec'
+
 # Steam::Browser::HtmlUnit::Drb::Service.daemonize and sleep(0.25)
 browser = Steam::Browser::HtmlUnit.new #(:drb => true)
 World do
@@ -16,10 +20,12 @@ end
 Before do
   ActiveRecord::Base.send(:subclasses).each do |model|
     connection = model.connection
-    if connection.instance_variable_get(:@config)[:adapter] == 'mysql'
-      connection.execute("TRUNCATE #{model.table_name}")
-    else
-      connection.execute("DELETE FROM #{model.table_name}")
+    if model.table_exists?
+      if connection.instance_variable_get(:@config)[:adapter] == 'mysql'
+        connection.execute("TRUNCATE #{model.table_name}")
+      else
+        connection.execute("DELETE FROM #{model.table_name}")
+      end
     end
   end
 end
